// src/navigators-types.ts
export type NavigatorKind = 'OnboarderNavigator' | 'ERC20TributeNavigator' | 'NFTGatedNavigator' | 'SignalNavigator' | 'TimelockNavigator' | 'VestingNavigator' | 'BudgetNavigator' | 'SubscriptionNavigator';